64 bit fixes, httpd work, spawn test

+++ libsxe/lib-sxe-pool-tcp/test/test-sxe-pool-tcp.c
@@ -75,7 +75,8 @@ test_event_client_read(SXE * this, int length)
 {
     SXEE81I("test_event_client_read(length=%d)", length);
     SXEA10I(length != 0, "test_event_client_read called with length == 0");
-    tap_ev_push(__func__, 4, "this", this, "user_data", SXE_USER_DATA(this), "length", length, "buf", tap_dup(SXE_BUF(this), SXE_BUF_USED(this)));
+    tap_ev_push(__func__, 4, "this", this, "user_data", SXE_USER_DATA(this), "length", length, "buf",
+                tap_dup(SXE_BUF(this), SXE_BUF_USED(this)));
     SXE_BUF_CLEAR(this);
     SXER80I("return");
 }
@@ -140,9 +141,9 @@ test_expect_server_read(const char * expected_buf, unsigned expected_length)
 
     SXEE63("%s(expected_buf=%p, expected_length=%d)", __func__, expected_buf, expected_length);
 
-    ok((ev = test_next_server_event()) != NULL,                      "Event occurred");
-    is_eq(tap_ev_identifier(ev), "test_event_server_read",            "Got a server read event");
-    is((unsigned)tap_ev_arg(ev, "length"), expected_length,           "Server received %u characters", expected_length);
+    ok((ev = test_next_server_event()) != NULL,             "Event occurred");
+    is_eq(tap_ev_identifier(ev), "test_event_server_read",  "Got a server read event");
+    is((uintptr_t)tap_ev_arg(ev, "length"), expected_length, "Server received %u characters", expected_length);
     this = (SXE *)(long)tap_ev_arg(ev, "this");
 
     /* If the expected buffer contains a string that is shorter than the expected length, just make sure the received buffer
@@ -150,7 +151,7 @@ test_expect_server_read(const char * expected_buf, unsigned expected_length)
      */
     if (strlen(expected_buf) < expected_length) {
         is_strncmp(tap_ev_arg(ev, "buf"), expected_buf, strlen(expected_buf), "Server received message matched /%s%.*s/",
-                   expected_buf, expected_length - strlen(expected_buf), "................");
+                   expected_buf, (int)(expected_length - strlen(expected_buf)), "................");
     }
     else {
         is_strncmp(tap_ev_arg(ev, "buf"), expected_buf, strlen(expected_buf), "Server received '%s'", expected_buf);
@@ -211,15 +212,15 @@ test_case_happy_path(void)
     this = test_expect_server_read("Hello", 5);
     sxe_write(this, "There", 5);
     is_eq(tap_ev_identifier(ev = test_next_client_event()), "test_event_client_read", "1st client read");
-    is(tap_ev_arg(ev, "length"), 5,                        "Five bytes of data received on the tcp pool object");
-    is_strncmp(tap_ev_arg(ev, "buf"), "There", 5,          "Received 'There' on the tcp pool object");
+    is((unsigned)(uintptr_t)tap_ev_arg(ev, "length"), 5,    "Five bytes of data received on the tcp pool object");
+    is_strncmp(tap_ev_arg(ev, "buf"), "There", 5,           "Received 'There' on the tcp pool object");
 
     /* Two sends at a time */
 
     test_check_ready_to_write(pool, 1, __func__, __LINE__);
-    is(sxe_pool_tcp_write(pool, "Hello2", 6, buf), SXE_RETURN_OK,    "Successfully wrote 'Hello2' to the tcp pool object");
+    is(sxe_pool_tcp_write(pool, "Hello2", 6, buf), SXE_RETURN_OK, "Successfully wrote 'Hello2' to the tcp pool object");
     test_check_ready_to_write(pool, 1, __func__, __LINE__);
-    is(sxe_pool_tcp_write(pool, "Hello3", 6, buf), SXE_RETURN_OK,    "Successfully wrote 'Hello3' to the tcp pool object");
+    is(sxe_pool_tcp_write(pool, "Hello3", 6, buf), SXE_RETURN_OK, "Successfully wrote 'Hello3' to the tcp pool object");
 
     /* Two responses at a time */
 
@@ -230,8 +231,8 @@ test_case_happy_path(void)
 
     for (i = 0; i < 2; i++) {
         is_eq(tap_ev_identifier(ev = test_next_client_event()), "test_event_client_read", "2nd or 3rd client read");
-        is(tap_ev_arg(ev, "length"), 3,                        "3 bytes of data received on the tcp pool object");
-        is_strncmp(tap_ev_arg(ev, "buf"), "xxx", 3,            "Received 'xxx' on the tcp pool object");
+        is((unsigned)(uintptr_t)tap_ev_arg(ev, "length"), 3,    "3 bytes of data received on the tcp pool object");
+        is_strncmp(tap_ev_arg(ev, "buf"), "xxx", 3,             "Received 'xxx' on the tcp pool object");
     }
 
     /* Shut 'er down */
